Updating to use sass modules (#12)
This makes some minor improvements in appearance.
It now uses sass modules instead of @import to eliminate warnings.

The theme switcher was removed, only the light-theme is supported, and some responsive layout was lost during the transition to modules. Every attempt was made to be backward compatible with any light-theme overrides. However, variables declared in _sass/_overrides.scss are not visible to _sass/_overrides-light.scss.

If someone has more sass experience, I'm open to suggestions for improving this.